requirements.txt
matplotlib
fastapi
uvicorn
tfx==1.15.0
python-snappy
numpy 
pandas
kfp==2.4.0
sympy
scikit-learn==1.5.0
joblib==1.4.2
imbalanced-learn==0.12.3
google-cloud-aiplatform
google-cloud-pipeline-components==2.6.0
keras_tuner==1.4.4
tensorflow_transform
tensorflow==2.15.1
tensorflow-data-validation==1.15.1
tensorflow_metadata